Add 'expiring' criteria
This also adds a field to mako_notification to track the requested
timeout, as opposed to what we decided to give it. This was required for
the value to be present at the time criteria are applied.

Probably the most useful thing you can do with this is set
ignore-timeout + default-timeout to prevent anything from sending you
permanent notifications.