Fix POSIX feature test macros
This commit applies fixes for the following:
pool-buffer.c: mkstemp is in POSIX.1-2008, thus requires _POSIX_C_SOURCE
to be 200809L or higher, and _X_OPEN_SOURCE 500 does not provide that.
event-loop.c: TFD_CLOEXEC is defined as O_CLOEXEC, which also appeared
in POSIX.1-2008, thus _POSIX_C_SOURCE 199309L is not enough.