Remove TODO about dbus_watch_handle error handling
A call to dbus_watch_handle fails if there is not enough memory. In such
case, the "application may spin in a busy loop on the file descriptor
until memory becomes available, but nothing more catastrophic should
happen" (from libdbus documentation). I think this is better, than
potentially blocking the handling of other fd events with a sleep.