Rework border drawing
This fixes the border to be visually outset from the background. Instead
of drawing the border first and trying to adjust the background to fit
inside it, we draw the background and progress, then draw the border on
top, replacing any overlapping color in case it's translucent.

This also changes how progress is drawn. Previously it was a nested
roundrect, which meant there would be a little bit of background leakage
at the corners due to the radius being calculated at a different width.
Now it is a normal rectangle, clipped to the background shape.