Add a history buffer for expired notifications
Notifications that have reached their timeout may
be added to a history buffer, displacing the
oldest stored notification if the buffer is full.

The buffer's size is controlled by the global
`max-history` setting. The `history` style
determines if a notification is saved to the
history buffer when it expires.

A new `RestoreNotification` dbus method restores
the most recent notification from the history buffer.

Restored notifications must be dismissed manually.

Closes: https://github.com/emersion/mako/issues/91