Add capabilities to use multiple surfaces
This is the solution after multiple iterations.

I tried to have it explicitly configured and creating the full superset
of all posibile surfaces.
The current version doesn't add much of a footprint in terms of code
and is very nice for the user.

This moves the configuration options `layer`, `output` and `anchor` into
style options and allows to configure them based on the criteria
selected for notifications.

Groups probably don't make sense when split accross surfaces and the
max-visible is currently taken from the first surface handled that is
put onto said surface.

Output hotplugging might cause weird behaviour. It might be interesting
to have a merge-automatic onto otputs if the prefered one is
unavailable, but that's left for later until there's a specific
use-case.