Include a systemd.service file
Including (and installing) these changes triggers calls to
`org.freedesktop.Notifications` to start a systemd user service with
mako.

This is pretty much how similar daemons (eg: dunst) behave. It allows
the service to be auto-started on demand, and also ties it to the
graphical session.

This also makes is trivial to restart `mako` like any other user service,
by just running `systemd --user restart mako`.

Users not running systemd (if that's supported) shouldn't be affected.