Reschedule frame if configuration is pending
If a surface is pending configuration, the state is dirty with no frame
callback and no configured flag. If a surface was destroyed in this
state, none of the rescheduling conditions would activate, while the
dirty flag would block all future frame scheduling.

Always schedule a new frame when dirty is set. Also replaces
need_reschedule with state->dirty directly.

Closes: https://github.com/emersion/mako/issues/189