Add support for body criteria
This adds the ability to match notifications based on notification
bodies, either using an exact string match or an approximate regex.

This functionality was present in the repo in a half-finished state for
a while, seemingly left until regex support - even as summaries gained
regex matching support. Thus, I've based this pretty much entirely on
summary matching.

I've done basic testing to ensure that this compiles and functions as
intended. However, since this is all pretty directly copied from the
summary code, any bugs in that are left uninvestigated, and changes are
limited to code regions where a find-all for "criteria" brought me.

The original motivation for this is laid out in issue
https://github.com/emersion/mako/issues/304 - mainly, filtering
messages for an application which does not differentiate based on
anything but notification body.

Hope this is useful!