Replace hidden_style with real criteria
Instead of being stored separately on the config, hidden is now
a criteria field. It is still special in that it is implicitly false on
all user-specified criteria. This prevents them from styling the
pseudo-notification that acts as the hidden placeholder, since this
would be very non-intuitive.

The placeholder is now a real notification object that gets created when
needed, so styling can make use of the same apply_each_criteria
mechanism as all other notifications. This allows hidden to be styled
per-surface. Validation is added to prevent styling hidden more
specifically than this, similar to max_visible.

Counting of hidden notifications has also been simplified because I had
to stare at it way too long to remember how it worked.