Store multiple modes
Keep a list of active modes.
This allows to mix multiple modes as specified in the configuration.

Allow to set the list to a single value and multiple values by a comma
separated list of modes. Add a new mode to the list by prefixing it with
a plus sign and remove a mode from the list by prefixing it with a minus
sign.