man: do not merge all makoctl mode invocations
In the manual page source file, the possible invocations of the makoctl
mode command (-s, -a, etc.) are listed in separate lines. In the
compiled manual page, they were merged to a single line, which made the
command invocation options hard to understand:
    mode mode -s <mode>... mode [-a mode]... [-r mode]...

Add explicit line breaks so that the options are listed in separate
lines in the final manual page as well. I have no idea if it is
semantically correct.

scdoc supports explicit line breaks since version 1.5.0 (2018-10-07),
mako seems to already require scdoc 1.9.7.