Add invoke-action <action>
Before this patch, it was impossible to send a different action
than `default` when clicking on a notification without loosing
access to the surface and thus the xdg_activation token.

With this patch, the user is now able to configure the action
to be sent when clicking on the notification while preserving
access to the xdg_activation token.

Example configuration:
```ini
[app-name="some-app-id" actionable]
on-button-left=invoke-action mail-reply-sender
```