fix(makoct): Ensure same notif is used across external call in `menu`
Consider the following:
The first notification has action `act1`. The user starts `makoctl menu`
with no id specified to get actions from this notifications. While dmenu
is open, a new notification appears. When the user confirms `act1` in
dmenu, this action previously got applied to this new notification,
which potentially has completely different actions/semantics.

Now the behaviour is that the action gets sent to the notification from
which it was obtained (this is consistent with behaviour when `-n` is
passed)