Fix use after free when closing notifications
This might be related to https://github.com/emersion/mako/issues/512, as
the problem I saw had similar symptoms. But at some point, I started to
see segfaults each time I closed a notification, regardless of whether I
(un)plugged monitors.

The crux of the issue is that closing a notification frees it, and in
some cases the notification was being used after being closed.