Check whether the notification supports action in `makoctl invoke`
When the action is not supported by the given notification, we return
an error condition and report this on the terminal.

It is useful to know whether the notification was successfully
invoked, for example, I have this in my `hyperland` configuration to
deal with the newest notification (similar to clicking on it):
```
bind = $MOD, period, exec, makoctl invoke || makoctl dismiss
```

An alternative approach for my use-case would be to have the
`InvokeAction` dismiss the notification when the `default` action is
sent to an action-less notification.  The chosen approach is more
versatile, as the user of `makoctl invoke` can decide what behavior to
chose.